Move RequireAtLeastOne to utils
